+++ app/scripts/bg/bg.js
@@ -1,6 +1,6 @@
 window._trackJs = {
   token: 'bd495185bd7643e3bc43fa62a30cec92',
-  enabled: false,
+  enabled: true,
   onError: function (payload) { 
     console.log('payload', payload)
     if (payload.message.indexOf('unknown') !== -1) {
@@ -257,7 +257,6 @@ var Bg = React.createClass({
   },
   componentDidMount(){
     this.listenTo(prefsStore, this.prefsChange);
-    this.queryTabs();
     this.querySessions();
   },
   prefsChange(e){
@@ -267,6 +266,7 @@ var Bg = React.createClass({
     this.setState({prefs: s.prefs});
     if (s.init) {
       this.attachListeners(s);
+      this.queryTabs(null, s.prefs);
       if (s.prefs.screenshot) {
         this.queryScreenshots();
       }
@@ -499,7 +499,7 @@ var Bg = React.createClass({
       } else if (msg.method === 'getTabs') {
         sendResponse({windows: this.state.windows, windowId: sender.tab.windowId});
       } else if (msg.method === 'queryTabs') {
-        this.queryTabs(true);
+        this.queryTabs(true, this.state.prefs);
       } else if (msg.method === 'getSessions') {
         sendResponse({sessions: this.state.sessions});
       } else if (msg.method === 'getScreenshots') {
@@ -511,11 +511,13 @@ var Bg = React.createClass({
         this.undoAction(this.state.windows[refWindow].tabs, this.state.chromeVersion);
       } else if (msg.method === 'getActions') {
         sendResponse({actions: this.state.actions});
+      } else if (msg.hasOwnProperty('scrollNav')) {
+        this.handleScrollNav(sender, msg.scrollNav === 'up' ? sender.tab.index + 1 : sender.tab.index - 1, msg.scrollNav);
       }
       return true;
     });
   },
-  formatTabs(tabs){
+  formatTabs(prefs, tabs){
     var blacklisted = [];
     _.each(tabs, (tab, tKey)=>{
       var urlMatch = tab.url.match(/^(?:https?:\/\/)?(?:[^@\n]+@)?(?:www\.)?([^:\/\n]+)/im);
@@ -526,8 +528,11 @@ var Bg = React.createClass({
       if (tab.url.indexOf('chrome://newtab/') !== -1) {
         blacklisted.push(tab.id);
       }
+      if (prefs.scrollNav && tab.url && tab.url.indexOf('chrome://') === -1 && tab.url.indexOf('chrome.google.com') === -1 && tab.url.indexOf('drive.google.com') === -1 && tab.url.indexOf('data:') === -1 && tab.url.indexOf('127.0.0.1') === -1 && tab.url.indexOf('localhost') === -1) {
+        this.executeScrollNav(tab.id);
+      }
     });
-    if (blacklisted.length > 0 && this.state.prefs && this.state.prefs.singleNewTab) {
+    if (blacklisted.length > 0 && prefs && prefs.singleNewTab) {
       var extraNewTabs = _.tail(blacklisted);
       console.log(extraNewTabs);
       for (let i = extraNewTabs.length - 1; i >= 0; i--) {
@@ -536,10 +541,10 @@ var Bg = React.createClass({
     }
     return tabs;
   },
-  queryTabs(send=null){
+  queryTabs(send=null, prefs){
     chrome.windows.getAll({populate: true}, (w)=>{
       _.each(w, (Window, wKey)=>{
-        Window.tabs = this.formatTabs(Window.tabs);
+        Window.tabs = this.formatTabs(prefs, Window.tabs);
       });
       this.setState({windows: w})
       if (send) {
@@ -709,7 +714,7 @@ var Bg = React.createClass({
       this.state.windows[refWindow].tabs.push(e);
     }
     this.state.windows[refWindow].tabs = _.orderBy(_.uniqBy(this.state.windows[refWindow].tabs, 'id'), ['pinned'], ['desc']);
-    this.state.windows[refWindow].tabs = this.formatTabs(this.state.windows[refWindow].tabs);
+    this.state.windows[refWindow].tabs = this.formatTabs(this.state.prefs, this.state.windows[refWindow].tabs);
     this.setState({windows: this.state.windows});
     setActionThrottled(this, 'create', e);
     synchronizeSession(this.state.sessions, this.state.prefs, this.state.windows);
@@ -747,6 +752,9 @@ var Bg = React.createClass({
         return;
       }
       var refTab = _.findIndex(this.state.windows[refWindow].tabs, {id: e.id});
+      if (refTab === -1) {
+        return;
+      }
       setActionThrottled(this, 'update', this.state.windows[refWindow].tabs[refTab], e);
       _.merge(e, {
         timeStamp: new Date(Date.now()).getTime()
@@ -791,6 +799,66 @@ var Bg = React.createClass({
         sendMsg({windows: this.state.windows, windowId: e.windowId});   
       }
     });
+  },
+  executeScrollNav(id){
+    chrome.tabs.executeScript(id, {
+      code: `
+        var wheelListener = (e)=>{
+          if (e.wheelDelta / 120 > 0) {
+            chrome.runtime.sendMessage({scrollNav: 'down'}); 
+          }
+          else {
+            chrome.runtime.sendMessage({scrollNav: 'up'}); 
+          }
+          e.preventDefault();
+        };
+        document.onmousemove = (e)=>{
+          if (e.y <= 20 || e.shiftKey) {
+            document.body.style.cursor = 'all-scroll';
+            window.addEventListener('mousewheel', wheelListener);
+          } else {
+            document.body.style.cursor = 'initial';
+            window.removeEventListener('mousewheel', wheelListener);
+          }
+        };
+      `
+    });
+
+  },
+  handleScrollNav(sender, index, direction){
+    if (this.state.prefs.scrollNav) {
+      var findTab = (_index)=>{
+        var refWindow = _.findIndex(this.state.windows, {id: sender.tab.windowId});
+        if (refWindow === -1) {
+          return;
+        }
+        var refTab = _.findIndex(this.state.windows[refWindow].tabs, {index: _index});
+        if (refTab === -1) {
+          var lastTab = _.last(this.state.windows[refWindow].tabs);
+          if (index === -1) {
+            findTab(lastTab.index);
+          } else if (index > lastTab.index) {
+            findTab(0);
+          }
+          return;
+        }
+        var tab = this.state.windows[refWindow].tabs[refTab];
+        if ((tab.url.indexOf('chrome://') !== -1 || tab.url.indexOf('chrome.google.com') !== -1|| tab.url.indexOf('127.0.0.1') !== -1 || tab.url.indexOf('localhost') !== -1 || tab.url.indexOf('drive.google.com') !== -1 || tab.url.indexOf('data:') !== -1) 
+          && tab.url.indexOf('chrome://newtab/') === -1
+          || tab.status === 'loading'
+          || tab.discarded) {
+          if (direction === 'up') {
+            findTab(++_index);
+          } else {
+            findTab(--_index);
+          }
+        } else {
+          console.log('scrollNav trigger');
+          chrome.tabs.update(this.state.windows[refWindow].tabs[refTab].id, {active: true});
+        }
+      };
+      findTab(index); 
+    }
   },
   render(){
     console.log('BG STATE: ',this.state);