+++ app/scripts/components/root.js
@@ -1,6 +1,6 @@
 window._trackJs = {
   token: 'bd495185bd7643e3bc43fa62a30cec92',
-  enabled: false,
+  enabled: true,
   onError: function (payload) { return true; },
   version: "",
   callback: {
@@ -241,6 +241,9 @@ var Root = React.createClass({
     var p = this.props;
     var stateUpdate = {};
     var sUChange = false;
+    if (nP.s.prefs.scrollNav !== p.s.prefs.scrollNav) {
+      this.handleMouseListeners(nP);
+    }
     if (nP.s.modeKey !== p.s.modeKey && nP.s.prefs.mode === p.s.prefs.mode) {
       if (nP.s.search.length > 0) {
         stateUpdate[nP.s.modeKey] = utils.searchChange(nP, nP.s[nP.s.modeKey]);
@@ -295,6 +298,46 @@ var Root = React.createClass({
         });
       }
     });
+    this.handleMouseListeners(p);
+  },
+  handleMouseListeners(p){
+    var wheelListener = (e)=>{
+      if (e.wheelDelta / 120 > 0) {
+        chrome.runtime.sendMessage({scrollNav: 'down'}); 
+      }
+      else {
+        chrome.runtime.sendMessage({scrollNav: 'up'}); 
+      }
+      e.preventDefault();
+    };
+    document.onmousemove = handleMouseMove;
+    function handleMouseMove(e) {
+      if (p.s.prefs.scrollNav) {
+        if (e.y <= 20 || e.shiftKey) {
+          document.body.style.cursor = 'all-scroll';
+          window.addEventListener('mousewheel', wheelListener);
+        } else {
+          document.body.style.cursor = 'initial';
+          window.removeEventListener('mousewheel', wheelListener);
+        }
+      } else {
+        window.removeEventListener('mousewheel', wheelListener);
+      }
+      utilityStore.set_cursor({
+        page: {
+          x: e.pageX,
+          y: e.pageY
+        },
+        offset: {
+          x: e.offsetX,
+          y: e.offsetY,
+        },
+        keys: {
+          ctrl: e.ctrlKey,
+          shift: e.shiftKey
+        }
+      });
+    }
   },
   faviconsChange(e){
     this.setState({topLoad: true});