+++ app/scripts/components/stores/prefs.js
@@ -35,7 +35,8 @@ var prefsStore = Reflux.createStore({
       theme: 9001,
       wallpaper: null,
       tooltip: true,
-      alerts: true
+      alerts: true,
+      scrollNav: false
     };
     var getPrefs = new Promise((resolve, reject)=>{
       chrome.storage.sync.get('preferences', (prefs)=>{
@@ -89,7 +90,8 @@ var prefsStore = Reflux.createStore({
         syncedSession: prefs.syncedSession,
         wallpaper: prefs.wallpaper,
         tooltip: prefs.tooltip,
-        alerts: prefs.alerts
+        alerts: prefs.alerts,
+        scrollNav: prefs.scrollNav
       };
       if (typeof this.prefs.autoDiscard === 'undefined') {
         this.prefs.autoDiscard = false;
@@ -144,6 +146,9 @@ var prefsStore = Reflux.createStore({
       }
       if (typeof this.prefs.alerts === 'undefined') {
         this.prefs.alerts = true;
+      }
+      if (typeof this.prefs.scrollNav === 'undefined') {
+        this.prefs.scrollNav = false;
       }
       console.log('load prefs: ', prefs, this.prefs);
       this.trigger(this.prefs);